Fix smoothing bug
The code to smooth the default Adafruit glyphs used an 'else if' where
just an 'if' should have been used as it is possible for a set pixel to
be the top of both a / diagonal and a \ diagonal. With this bug, the
top of a glyph like 'A' would only smooth the diagonal on the left and
not the one on the right.