Change layout of font array
I wanted to be able to look directly at the glyphs representing the
main printable characters in the default font so I reformatted them to
place each common glyph on its own line with a trailing comment showing
which character it represented. This made it easier to see how to
smooth each the ones that the diagonal only smoothing didn't completely
address.