Smooth more of the default font glyphs
The diagonal smoothing that I added earlier helps a lot to smooth out
the default Adafruit glyphs but it still left jaggies on some of the
glyphs. This commit addresses several of those cases like '0', '1',
'3', '4', 'M', 'N', etc. It adds the g_hints array to Adafruit_GFX.cpp
which describes extra triangles to be added to specific glyphs when
they are scaled up. This array only contains entries for the glyphs
which actually need additional smoothing help.

I also moved the smoothing version of the drawChar() method out of the
ColorMemLCD class and placed it into the Adafruit_GFX class instead.
This will make it easier for me to reuse this font smoothing code in
future projects.